Fix warning under MSVC
Use an explicit static_cast in coverting a `size_t` to a `socklen_t`.

Before this fix, the warning
```
warning C4267: '=' : conversion from 'size_t' to 'socklen_t', possible loss of data
```
is produced, and now it's suppressed. This is problematic because this also produces a warning in code using SRT, which could have warnings set as errors.