Made WinSock non-blocking #383
Previously we set the minimum recv timeout to 1ms on Windows, but there is an undocumented minimum receive timeout of 500ms, according to some sources online. As a result, some packets were lost for srt, but were still received by wireshark. This commit sets the socket to non-blocking via ioctl