Fix Windows installers
The Windows installer provides the SRT libraries for all possible
architectures. However, SRT needs OpenSSL and OpenSSL for Windows
cannot install its Arm64 libraries on Intel systems. More precisely,
on Arm64 systems, we may install OpenSSL libraries for x86, x64 and
Arm64. However, on Intel systems, OpenSSL libraries can be installed
for x86, x64 only. This is an unfortunate limitation of the OpenSSL
installer for Windows.

As a consequence, when building the libsrt installer in a Windows
Intel system, the installer now contains the x86 and x64 libraries
only. To build a libsrt installer with the binaries for the three
architectures, the build must be performed on a Windows Arm64 system.

This is not easy to achieve on automated CI systems where Windows
systems are usually Intel only. Manually, the easiest way to get
a Windows Arm64 is to install a virtual machine on a recent Mac.